Added settings manager to main menu scene. UI controls now load their values from the settings manager.
Also, with the help of new main menu script, pressing esc to go back to
previous menu works in the main menu scene too.

Settings manager is a static singleton object which is passed over to
the master scene. Recommend removing it (and the game manager) from the
master scene.